// lib/types/pdfStyle.ts
export type PdfStyle = {
    title: { fontSize: number };
    h1: { fontSize: number };
    h2: { fontSize: number };
    h3: { fontSize: number };
    h4: { fontSize: number };
    h5: { fontSize: number };
    h6: { fontSize: number; bold: boolean };
    p: { fontSize: number };
    small: { fontSize: number };
    sub: { fontSize: number };
    right: { alignment: string };
    left: { alignment: string };
    topMargin1: { margin: [number, number, number, number] };
    topMargin2: { margin: [number, number, number, number] };
    topMargin3: { margin: [number, number, number, number] };
    topMargin4: { margin: [number, number, number, number] };
    topMarginRegular: { margin: [number, number, number, number] };
    tableMargin: { margin: [number, number, number, number] };
    b: { bold: boolean };
    i: { italics: boolean };
    primary: { color: string };
    alternate: { color: string };
    gray: { color: string };
    lightGray: { color: string };
    darkGray: { color: string };
    red: { color: string };
    blue: { color: string };
    mono: { font: string; fontSize: number };
    monoSub: { font: string; fontSize: number };
};